Updated CLI application and setup script
